# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Song library",
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/info": {
            "get": {
                "description": "Возвращает информацию о песне",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получить информацию о песне по группе и названии песни",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название группы",
                        "name": "group",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название песни",
                        "name": "song",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о песне",
                        "schema": {
                            "$ref": "#/definitions/model.SongDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос, отсутствуют обязательные параметры",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Песня не найдена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавление песни с приведенными параметрами.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Создание новой песни",
                "parameters": [
                    {
                        "description": "Информация о песне для создания",
                        "name": "songDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SongDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о созданной песни",
                        "schema": {
                            "$ref": "#/definitions/model.SongDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос - неправильные входные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновлеят данные песни по группе и названию песни. Обновлеят только переданные данные.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Обновляет данные песни",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название песни для обновления данных",
                        "name": "song",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название группы для обновления данных",
                        "name": "group",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Новые данные. Будут обновлены только переданные параметры!",
                        "name": "newInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SongDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные песни",
                        "schema": {
                            "$ref": "#/definitions/model.SongDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос, отсутсвуют необходимые данные о песне (group, song)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/info/all": {
            "get": {
                "description": "Возвращает список песен с пагинацией (limit и offset).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получить информацию о всех песнях",
                "parameters": [
                    {
                        "maximum": 50,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Максимальное количество возвращаемых песен",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "maximum": 25,
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Количетство пропускаемых песен",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список песен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SongDetailResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "No songs found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/info/delete": {
            "delete": {
                "description": "Удаляет запись о песне по группе и названию песни.",
                "tags": [
                    "Songs"
                ],
                "summary": "Удаляет запись о песне",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название группы для удаления",
                        "name": "group",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название песни для удаления",
                        "name": "song",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Песня успешна удалена",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос, отсуствуют необходимые данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Не найдено - такой песни нет в базе данных",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/info/filter": {
            "post": {
                "description": "Возвращает список песен с пагинацией (limit и offset).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получить песни по фильтру",
                "parameters": [
                    {
                        "maximum": 50,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Максимальное количество возвращаемых песен",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "maximum": 25,
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Количетство пропускаемых песен",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "description": "Информация о песни, по которой будет происходить фильтрация",
                        "name": "songDetail",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SongDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список песен по фильтру",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SongDetailResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос - неправильные входные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Не найдено - нет песен удовлетворяющих фильтру",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервиса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/info/text": {
            "get": {
                "description": "Возвращает текст песни на основе названия песни и группы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Songs"
                ],
                "summary": "Получить текст песни",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя группы",
                        "name": "group",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название песни",
                        "name": "song",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Текст песни",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный необходимо передать group, song как query param",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Текст песни не добавлен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.SongDetailRequest": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string",
                    "example": "Muse"
                },
                "link": {
                    "type": "string",
                    "example": "https://www.youtube.com/watch?v=Xsp3_a-PMTw"
                },
                "release_date": {
                    "type": "string",
                    "example": "14.03.2024"
                },
                "song": {
                    "type": "string",
                    "example": "Supermassive Black Hole"
                },
                "text": {
                    "type": "string",
                    "example": "Ooh baby, don't you know I suffer?\nOoh baby, can you hear me moan?\nYou caught me under false pretenses\nHow long before you let me go?\n\nOoh\nYou set my soul alight\nOoh\nYou set my soul alight"
                }
            }
        },
        "model.SongDetailResponse": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string",
                    "example": "Muse"
                },
                "link": {
                    "type": "string",
                    "example": "https://www.youtube.com/watch?v=Xsp3_a-PMTw"
                },
                "release_date": {
                    "type": "string",
                    "example": "14.03.2024"
                },
                "song": {
                    "type": "string",
                    "example": "Supermassive Black Hole"
                },
                "text": {
                    "type": "string",
                    "example": "Ooh baby, don't you know I suffer?\nOoh baby, can you hear me moan?\nYou caught me under false pretenses\nHow long before you let me go?\n\nOoh\nYou set my soul alight\nOoh\nYou set my soul alight"
                }
            }
        }
    }
}